Refactor : Chat 컴포넌트의 인피니트 스크롤, 소켓 통신 로직을 각각의 커스텀 훅으로 분리
